# util/deephash: simplify hasher.hashMap