@dsnet 2021-07-23
Approximately 35 lines and 4.5KiB (for compiled x64 output) reduction in complexity.